email-processing/agent/requirements.txt
fastapi[All]
azure-identity
msgraph-sdk
msal